{wrapper, cilpp, cccppp}: refactor command-line processing to be more flexible
Previously the command-line processing was all in one big function and
existed only to allow preprocessing into a temporary file and then
feeding that file onwards... it was essentially cilpp-specific. Now it
is split into multiple functions that can be used in various ways... the
client composes them as they see fit.