wrapper: various cleanups, including unbreaking the use of -MD
The code was confused about the meanings of -M, -MM and -MD. It was also
confused about where the form "-MD <file>" appears: the answer is "only
in cc1 commands". This is a nasty irregularity... the wrapper script
code is now more explicit about when it is dealing with cpp commands,
when with cc1 commands and when with cc commands, and also when just with
"options" (lacking the executable name) rather than full commands.

I suspect cccppp is now broken; I haven't tried to keep it working, as
it needs more serious attention.