Add `cabsprint` tool
We need a way of investigating the cabs representation of a program --- `cilprint` etc will only print the cil model (i.e. processed by cabs2cil).

This adds `cabsprint`, which prints the cabs representation of an input program and otherwise behaves similarly to cilprint and cilpp.

It also adds a `-cabs-structure` flag which only works with `cabsprint`. `-cabs-structure` toggles the printing of comments that indicate the structure of the cabs representation (rather than just plain C).